Externalize `ModuleHandler` subclasses
