+++ src/rookify/modules/create_cluster/main.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 
-import kubernetes
 from typing import Any
+from ..exception import ModuleException
 from ..machine import Machine
-from ..module import ModuleHandler, ModuleException
+from ..module import ModuleHandler
 
 
 class CreateClusterHandler(ModuleHandler):
@@ -16,7 +16,7 @@ class CreateClusterHandler(ModuleHandler):
 
     @property
     def __mon_placement_label(self) -> str:
-        return (
+        return (  # type: ignore
             self._config["rook"]["cluster"]["mon_placement_label"]
             if "mon_placement_label" in self._config["rook"]["cluster"]
             else f"placement-{self._config["rook"]["cluster"]["name"]}-mon"
@@ -24,7 +24,7 @@ class CreateClusterHandler(ModuleHandler):
 
     @property
     def __mgr_placement_label(self) -> str:
-        return (
+        return (  # type: ignore
             self._config["rook"]["cluster"]["mgr_placement_label"]
             if "mgr_placement_label" in self._config["rook"]["cluster"]
             else f"placement-{self._config["rook"]["cluster"]["name"]}-mgr"
@@ -104,13 +104,9 @@ class CreateClusterHandler(ModuleHandler):
 
         self.k8s.crd_api_apply(cluster_definition)
 
-        cluster_name = self._config["rook"]["cluster"]["name"]
-
         # Wait for CephCluster to get into Progressing phase
-        result = None
-        watcher = kubernetes.watch.Watch()
-
-        stream = watcher.stream(
+        result = self.k8s.watch_events(
+            self._watch_cluster_phase_callback,
             self.k8s.custom_objects_api.list_namespaced_custom_object,
             "ceph.rook.io",
             "v1",
@@ -119,23 +115,21 @@ class CreateClusterHandler(ModuleHandler):
             timeout_seconds=60,
         )
 
-        for event in stream:
-            event_object = event["object"]
-
-            if event_object["metadata"]["name"] != cluster_name:
-                continue
-
-            try:
-                if event_object["status"]["phase"] == "Progressing":
-                    result = event_object
-                    break
-            except KeyError:
-                pass
-
-        watcher.stop()
-
-        if result == None:
+        if result is None:
             raise ModuleException("CephCluster did not come up")
+
+    def _watch_cluster_phase_callback(self, event_object: Any) -> Any:
+        try:
+            if (
+                event_object["metadata"]["name"]
+                == self._config["rook"]["cluster"]["name"]
+                and event_object["status"]["phase"] == "Progressing"
+            ):
+                return event_object
+        except KeyError:
+            pass
+
+        return None
 
     @staticmethod
     def register_preflight_state(
