+++ src/rookify/modules/exception.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+
+class ModuleException(Exception):
+    pass
