+++ tests/modules/test_example.py
@@ -3,8 +3,8 @@
 import pytest
 
 from rookify.modules.example.main import ExampleHandler
+from rookify.modules.exception import ModuleException
 from rookify.modules.machine import Machine
-from rookify.modules.module import ModuleException
 
 
 def test_preflight() -> None:
